Created Comparator
Comparator class should be used for comparing the performance between
serial and parallel version of the project. It has a method for loading
all test cases (lists of numbers to be used to generate expected solution)
and methods for running the Computer (class for generating each expression)
on every single test case for each of the versions of project. Also has
a function for serializing each result into csv, so it could be easily
later displayed in Excel or Google sheets.